tests/wpt/css-tests/css-flexbox-1_dev/xhtml1/order_value.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head>
    <title>CSS Flexible Box Test: order_check</title>
    <link href="http://www.intel.com" rel="author" title="Intel" />
    <link href="http://www.w3.org/TR/css-flexbox-1/#propdef-order" rel="help" />
    <link href="support/test-style.css" rel="stylesheet" />
    <meta content="" name="flags" />
    <meta content="Check if the web engine can indentify order property." name="assert" />
    <style>
      #test{
	height: 200px;
 	width: 200px;
        display: flex;
      }
      div#test02{
	order: -1;
      }
    </style>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <div id="test"><div id="test01">1</div><div id="test02">2</div><div id="test03">3</div></div>
    <script type="text/javascript">
        var t = async_test(document.title, {timeout:500});
        t.step(function () {
            assert_equals(document.getElementById("test01").offsetTop, document.getElementById("test02").offsetTop, "Rectangle 1 and 2 have the same offsetTop value");
            assert_equals((document.getElementById("test02").offsetLeft &gt;= document.getElementById("test01").offsetLeft), false, "Rectangle 2 have a smaller offsetLeft value than 1.");
        });
        t.done();
    </script>
  

</body></html>
